backend/requirements.txt
fastapi==0.110
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-jose==3.3.0
python-dotenv==1.0.0
